requirements.txt
typer==0.9.0
rich==13.7.0
platformdirs==4.0.0
pyinstaller==6.3.0
pandas>=2.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
matplotlib>=3.8.0
python-dateutil
numpy>=1.26.0
plotext>=5.2.0
ofxparse>=0.21
pdfplumber>=0.10.0
scikit-learn>=1.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
dropbox>=11.36.2
keyring>=24.2.0
cryptography>=41.0.0
gitpython>=3.1.40